// src/types/mutations.ts
export enum MUTATIONS {
    /* eslint-disable no-unused-vars */
    SET_LOCATIONS = 'setLocation',
    SET_BANNERS = 'setBanners',
    SET_LATEST_HOTEL = 'setLatestHotel',
    SET_FEATURES_HOTEL = 'setFeaturedHotel',
    SET_SEARCH_RESULT = 'setSearchResult',
    SET_PREVIEW_HOTEL_LIST = 'setPreviewHotelList',
    CLEAR_SEARCH_RESULT = 'clearSearchResult',
    FETCH_BANNERS = 'fetchBanners',
    SET_LOADING = 'setLoading',
    SET_SOME_HOTEL_INFO = 'setSomeHotelInfo'
}